+++ src/system/Zikula/Module/SettingsModule/Api/AdminApi.php
@@ -17,6 +17,7 @@ namespace Zikula\Module\SettingsModule\Api;
 use ZLanguage;
 use ModUtil;
 use SecurityUtil;
+use System;
 use Zikula_View_Theme;
 use Zikula_View;
 
@@ -53,6 +54,21 @@ class AdminApi extends \Zikula_AbstractApi
         Zikula_View_Theme::getInstance()->clear_cssjscombinecache();
         Zikula_View::getInstance()->clear_all_cache();
         Zikula_View::getInstance()->clear_compiled();
+
+        return true;
+    }
+
+    /**
+     * Resets the startmodule to static frontpage.
+     *
+     * @return boolean true.
+     */
+    public function resetStartModule()
+    {
+        System::setVar('startpage', '');
+        System::setVar('starttype', '');
+        System::setVar('startfunc', '');
+        System::setVar('startargs', '');
 
         return true;
     }